DSL_assignment9
A palindrome is a string of character that„s the same forward and backward.
Typically, punctuation, capitalization, and spaces are ignored. For example, “Poor
Dan is in a droop” is a palindrome, as can be seen by examining the characters
“poor Danish is in a droop” and observing that they are the same forward and
backward. One way to check for a palindrome is to reverse the characters in the
string and then compare with them the original-in a palindrome, the sequence will

be identical. Write C++ program with functions-
a) To print original string followed by reversed string using stack

b) To check whether given string is palindrome or not